+++ CHANGELOG.rst
@@ -22,6 +22,10 @@ Changelog
 
 * Add ``--target-version`` option to specify target Django version.
 
+* Add Django 3.1+ fixer to migrate ``{% trans %}`` to ``{% translate %}`` and ``{% blocktrans %}`` / ``{% endblocktrans %}`` to ``{% blocktranslate %}`` / ``{% endblocktranslate %}``.
+
+  `PR #52 <https://github.com/adamchainz/djade/pull/52>`__.
+
 * Add Django 3.1+ fixer to migrate ``{% ifequal %}`` / ``{% endifequal %}`` and ``{% ifnotequal %}`` / ``{% endifnotequal %}`` to ``{% if %}`` / ``{% endif %}``.
 
   `PR #35 <https://github.com/adamchainz/djade/pull/35>`__.